Extract TablesConfiguration from InstanceConfiguration
